+++ src/main/java/org/folio/circulation/domain/MoveRequestProcessAdapter.java
@@ -3,7 +3,7 @@ package org.folio.circulation.domain;
 import java.util.concurrent.CompletableFuture;
 
 import org.folio.circulation.infrastructure.storage.inventory.ItemRepository;
-import org.folio.circulation.infrastructure.storage.LoanRepository;
+import org.folio.circulation.infrastructure.storage.loans.LoanRepository;
 import org.folio.circulation.infrastructure.storage.requests.RequestQueueRepository;
 import org.folio.circulation.infrastructure.storage.requests.RequestRepository;
 import org.folio.circulation.support.Result;
