+++ src/main/java/org/folio/circulation/domain/OverdueFineCalculatorService.java
@@ -16,12 +16,12 @@ import java.util.function.Predicate;
 import org.apache.commons.lang3.ObjectUtils;
 import org.folio.circulation.infrastructure.storage.notices.ScheduledNoticesRepository;
 import org.apache.commons.lang3.StringUtils;
-import org.folio.circulation.infrastructure.storage.LoanPolicyRepository;
-import org.folio.circulation.infrastructure.storage.LostItemPolicyRepository;
+import org.folio.circulation.infrastructure.storage.loans.LoanPolicyRepository;
+import org.folio.circulation.infrastructure.storage.loans.LostItemPolicyRepository;
 import org.folio.circulation.domain.policy.OverdueFineCalculationParameters;
 import org.folio.circulation.domain.policy.OverdueFineInterval;
 import org.folio.circulation.domain.policy.OverdueFinePolicy;
-import org.folio.circulation.infrastructure.storage.OverdueFinePolicyRepository;
+import org.folio.circulation.infrastructure.storage.loans.OverdueFinePolicyRepository;
 import org.folio.circulation.domain.representations.StoredAccount;
 import org.folio.circulation.domain.representations.StoredFeeFineAction;
 import org.folio.circulation.infrastructure.storage.feesandfines.AccountRepository;
