+++ src/main/java/org/folio/circulation/domain/validation/ItemLimitValidator.java
@@ -12,7 +12,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.folio.circulation.domain.Item;
 import org.folio.circulation.domain.Loan;
 import org.folio.circulation.domain.LoanAndRelatedRecords;
-import org.folio.circulation.infrastructure.storage.LoanRepository;
+import org.folio.circulation.infrastructure.storage.loans.LoanRepository;
 import org.folio.circulation.rules.AppliedRuleConditions;
 import org.folio.circulation.support.Result;
 import org.folio.circulation.support.ValidationErrorFailure;
