+++ src/main/java/org/folio/circulation/infrastructure/storage/CirculationPolicyRepository.java
@@ -30,7 +30,7 @@ public abstract class CirculationPolicyRepository<T> {
   private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
   private final CirculationRulesClient circulationRulesClient;
-  final CollectionResourceClient policyStorageClient;
+  protected final CollectionResourceClient policyStorageClient;
 
   protected CirculationPolicyRepository(
     CirculationRulesClient circulationRulesClient,
