+++ src/main/java/org/folio/circulation/resources/LoanCollectionResource.java
@@ -14,7 +14,7 @@ import org.folio.circulation.infrastructure.storage.feesandfines.AccountReposito
 import org.folio.circulation.domain.Item;
 import org.folio.circulation.domain.Loan;
 import org.folio.circulation.domain.LoanAndRelatedRecords;
-import org.folio.circulation.infrastructure.storage.LoanRepository;
+import org.folio.circulation.infrastructure.storage.loans.LoanRepository;
 import org.folio.circulation.domain.LoanRepresentation;
 import org.folio.circulation.domain.LoanService;
 import org.folio.circulation.infrastructure.storage.users.PatronGroupRepository;
@@ -26,9 +26,9 @@ import org.folio.circulation.domain.UpdateRequestQueue;
 import org.folio.circulation.domain.User;
 import org.folio.circulation.infrastructure.storage.users.UserRepository;
 import org.folio.circulation.domain.notice.schedule.DueDateScheduledNoticeService;
-import org.folio.circulation.infrastructure.storage.LoanPolicyRepository;
-import org.folio.circulation.infrastructure.storage.LostItemPolicyRepository;
-import org.folio.circulation.infrastructure.storage.OverdueFinePolicyRepository;
+import org.folio.circulation.infrastructure.storage.loans.LoanPolicyRepository;
+import org.folio.circulation.infrastructure.storage.loans.LostItemPolicyRepository;
+import org.folio.circulation.infrastructure.storage.loans.OverdueFinePolicyRepository;
 import org.folio.circulation.domain.validation.AlreadyCheckedOutValidator;
 import org.folio.circulation.domain.validation.ChangeDueDateValidator;
 import org.folio.circulation.domain.validation.ItemNotFoundValidator;
