+++ src/main/java/org/folio/circulation/resources/RequestFromRepresentationService.java
@@ -8,7 +8,7 @@ import static org.folio.circulation.support.Result.succeeded;
 import java.util.concurrent.CompletableFuture;
 
 import org.folio.circulation.domain.Loan;
-import org.folio.circulation.infrastructure.storage.LoanRepository;
+import org.folio.circulation.infrastructure.storage.loans.LoanRepository;
 import org.folio.circulation.domain.Request;
 import org.folio.circulation.domain.RequestAndRelatedRecords;
 import org.folio.circulation.infrastructure.storage.requests.RequestQueueRepository;
