+++ src/main/java/org/folio/circulation/resources/renewal/OverrideRenewalByBarcodeResource.java
@@ -5,7 +5,7 @@ import static org.folio.circulation.resources.renewal.RenewByBarcodeRequest.rene
 import java.util.concurrent.CompletableFuture;
 
 import org.folio.circulation.domain.Loan;
-import org.folio.circulation.infrastructure.storage.LoanRepository;
+import org.folio.circulation.infrastructure.storage.loans.LoanRepository;
 import org.folio.circulation.infrastructure.storage.users.UserRepository;
 import org.folio.circulation.storage.SingleOpenLoanByUserAndItemBarcodeFinder;
 import org.folio.circulation.infrastructure.storage.inventory.ItemRepository;
